+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.48.0] - 2020-10-07
+### Added
+- peer/hashring32: Support using an application header as the shard key, instead
+  of `transport.Request#ShardKey`
+- peer/hashring32: options NumReplicas and NumPeersEstimate are not private
+  anymore and can be used by consumer of the pkg.
+- observability: Add caller request TTL histogram in inbound middleware
+
+
 ## [1.47.2] - 2020-09-16
 ### Fixed
 - errors returned by `yarpcerrors.FromError` now behave properly when using
@@ -1281,6 +1290,7 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
+[1.48.0]: https://github.com/yarpc/yarpc-go/compare/v1.47.2...v1.48.0
 [1.47.2]: https://github.com/yarpc/yarpc-go/compare/v1.47.1...v1.47.2
 [1.47.1]: https://github.com/yarpc/yarpc-go/compare/v1.47.0...v1.47.1
 [1.47.0]: https://github.com/yarpc/yarpc-go/compare/v1.46.0...v1.47.0